Adds some Python related installation scripts.
Adds installation of python3 and pip3.
Adds installation of virtualenv and virtualenvwrapper in the system's pip3.
Adds installation of jupyter and tqdm in the user's pip3.
Adds a file containing powerline configuration for ipython.
Reworks some of the aliases.
Reworks update_env.sh.
Changes some of powerline's visuals in the zshrc.
Adds virtualenvwrapper plugin to oh-my-zsh in the zshrc.